# Improve helpful--key-sequence performance