After the success of https://github.com/Wilfred/helpful/pull/132, I thought it
might be fun to try to find more `helpful` performance improvements.
It turns out that `helpful--key-sequences` was a major hotspot.

The first two commits make minor tweaks in the body of
`helpful--key-sequences`. They wouldn't be of any consequence in a
typical function, but because they're in a hotspot, I think they do
make a difference.

The third commit is the big one. `helpful--key-sequences` gets mapped
over a long list of keymaps, and on every call it was generating a
list of global map keycodes for its symbol. But this list of keycodes
is the same every time, so there's no need to keep generating it.
Getting this list once and passing it into the map calls makes a huge
difference in performance in both speed and memory.

To measure, I used this function:

```
(defun helpful-benchmark ()
  (interactive)
  (benchmark 50 '(helpful-callable #'move-beginning-of-line)))
```

Just to be safe, I benchmarked before and after over the course of
several computer seances:

#### Before
```
  Elapsed time: 56.672061s (21.927753s in 160 GCs)
  Elapsed time: 58.963714s (23.069360s in 171 GCs)
  Elapsed time: 56.702608s (22.232472s in 155 GCs)
  Elapsed time: 55.452360s (19.637100s in 150 GCs)
```

#### After
```
  Elapsed time: 29.617945s (11.067265s in 72 GCs)
  Elapsed time: 26.197627s (8.385313s in 59 GCs)
  Elapsed time: 29.024914s (11.277491s in 78 GCs)
  Elapsed time: 28.599935s (10.094264s in 65 GCs)
```

So, it looks about a 50% speed improvement!

(Times were recorded on my crappy little Thinkpad.
`GNU Emacs 26.1.50 (build 1, x86_64-pc-linux-gnu, GTK+ Version 3.22.30) of 2018-06-25`)