@nickdrozd 2018-07-16
Added a test case to catch remapping