@nickdrozd 2018-07-16
Looks like that first commit was doing more work than I thought. Without it, we get:

`Elapsed time: 41.628086s (14.618057s in 95 GCs)`