@nickdrozd 2018-07-26
If there are no problems, can this be merged? I've been running it locally for a while with no issues AFAICT. The performance boost is noticeable.