@Wilfred 2018-07-28
Thanks :)

By the way, spending 35% of time in GC is quite a lot. We might benefit from let-binding `gc-cons-threshold` to a higher value somewhere.